pyproject.toml
[project]
name = "comics-mcp"
version = "0.1.0"
description = "LLM-oriented MCP server for querying comic character."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
  "requests>=2.32.0",
  "python-dotenv>=1.1.0"
]

[project.urls]
Homepage = "https://github.com/imikailoby/comics-mcp"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
